+++ WindowsFormsApp1/UserControl/PrirodnaDjubrivaUC.cs
@@ -109,20 +109,24 @@ namespace WindowsFormsApp1
         {
             try
             {
-
-
-
-                DialogResult Brisi;
-                Brisi = MessageBox.Show("Da li ste sigurni?", "Izbriši", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
-                if (Brisi == DialogResult.Yes)
+                if (textBox1.Text == "" || textBox2.Text == "" || comboBox1.Text == "" || comboBox2.Text == "" || dateTimePicker2.Text == "")
                 {
-                    string barKod = dataGridView1.SelectedRows[0].Cells[6].Value + string.Empty;
-                    var prirodnadjubriva = this.unit.PrirodnaDjubrivaa.GetPrirodnaDjubrivaBybarKod(int.Parse(barKod));
-                    this.unit.PrirodnaDjubrivaa.DeletePrirodnaDjubriva(prirodnadjubriva);
-                    this.unit.Complete();
-                    MessageBox.Show("Izbrisali ste hemikaliju.");
+                    MessageBox.Show("Izaberite artikal koji  želite da izbrišete ili popunite prazna polja!");
                 }
+                else
+                {
 
+                    DialogResult Brisi;
+                    Brisi = MessageBox.Show("Da li ste sigurni?", "Izbriši", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
+                    if (Brisi == DialogResult.Yes)
+                    {
+                        string barKod = dataGridView1.SelectedRows[0].Cells[6].Value + string.Empty;
+                        var prirodnadjubriva = this.unit.PrirodnaDjubrivaa.GetPrirodnaDjubrivaBybarKod(int.Parse(barKod));
+                        this.unit.PrirodnaDjubrivaa.DeletePrirodnaDjubriva(prirodnadjubriva);
+                        this.unit.Complete();
+                        MessageBox.Show("Izbrisali ste hemikaliju.");
+                    }
+                }
             }
             catch (Exception ex)
             {
