+++ WindowsFormsApp1/UserControl/ZaposleniUC.cs
@@ -28,7 +28,7 @@ namespace WindowsFormsApp1
 
         private void button1_Click(object sender, EventArgs e)
         {
-            if (textBox1.Text == "" || dateTimePicker2.Text == "" || textBox3.Text == "")
+            if (textBox1.Text == "" || dateTimePicker2.Text == "" || textBox3.Text == "" || textBox2.Text=="" || textBox4.Text=="" )
             {
                 MessageBox.Show("Morate popuniti sva polja!");
             }
@@ -46,7 +46,8 @@ namespace WindowsFormsApp1
                             // barKod = int.Parse(textBox3.Text), ovo je visak, barKod je autoIncrement //
 
                             datumRodjenja = dateTimePicker2.Value,
-
+                            KorisnickoIme=textBox2.Text,
+                            lozinka=textBox4.Text
 
                         };
                         this.unit.Zaposlenii.AddZaposleni(novi);
@@ -84,24 +85,49 @@ namespace WindowsFormsApp1
         {
             try
             {
-
-
-
-                DialogResult Brisi;
-                Brisi = MessageBox.Show("Da li ste sigurni?", "Izbriši", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
-                if (Brisi == DialogResult.Yes)
+                if (textBox1.Text == "" || dateTimePicker2.Text == "" || textBox3.Text == "" || textBox2.Text == "" )
                 {
-                    string korisnickoime = dataGridView1.SelectedRows[0].Cells[0].Value + string.Empty;
-                    var zaposleni = this.unit.Zaposlenii.GetZaposleniByKorisnickoIme(korisnickoime);
-                    this.unit.Zaposlenii.DeleteZaposleni(zaposleni);
-                    this.unit.Complete();
-                    MessageBox.Show("Izbrisali ste hemikaliju.");
+                    MessageBox.Show("Izaberite osobu koju želite da izbrišete ili unesite podatke!");
+                }
+
+                else
+                {
+                    DialogResult Brisi;
+                    Brisi = MessageBox.Show("Da li ste sigurni?", "Izbriši", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
+                    if (Brisi == DialogResult.Yes)
+                    {
+                        string korisnickoime = dataGridView1.SelectedRows[0].Cells[0].Value + string.Empty;
+                        var zaposleni = this.unit.Zaposlenii.GetZaposleniByKorisnickoIme(korisnickoime);
+                        this.unit.Zaposlenii.DeleteZaposleni(zaposleni);
+                        this.unit.Complete();
+                        MessageBox.Show("Izbrisali ste odabranu osobu.");
+                    }
                 }
 
             }
             catch (Exception ex)
             {
                 MessageBox.Show("Doslo je do greske, proverite unete podatke");
+            }
+        }
+
+        private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)
+        {
+            if (dataGridView1.SelectedRows.Count > 0)
+            {
+
+                string ime = dataGridView1.SelectedRows[0].Cells[2].Value + string.Empty;
+                string datumro = dataGridView1.SelectedRows[0].Cells[4].Value + string.Empty;
+                string prezime = dataGridView1.SelectedRows[0].Cells[3].Value + string.Empty;
+                string korisnickoime = dataGridView1.SelectedRows[0].Cells[0].Value + string.Empty;
+            
+
+                textBox1.Text = ime;
+                dateTimePicker2.Value = DateTime.Parse(datumro);
+                
+                textBox2.Text = korisnickoime;
+                textBox3.Text = prezime;
+
             }
         }
     }
